requirements.txt
tensorflow==2.18.0
numpy==1.26.2
opencv-python==4.10.0.84
Pillow==10.4.0
imgaug==0.4.0
matplotlib==3.9.2
scikit-learn==1.5.1
flask==3.0.3
Flask-RESTful==0.3.10
